image_manager: Check if tmp dir exists before remove
There was an error seen where the version manager core dumped when
it tried to remove the temporary directory after a manifest failure:

phosphor-version-software-manager[1264]: Error No manifest file
phosphor-version-software-manager[1264]: terminate called after throwing an
    instance of 'std::experimental::filesystem::v1::__cxx11::filesystem_error'
phosphor-version-software-manager[1264]:   what():  filesystem error: cannot
    remove all: No such file or directory [/tmp/images/imageKcwJFc]

To prevent the core dump, check that the directory exists before trying
to remove it. Systemd cleans the tmp directory every few days so it
could be possible the tmp directory is automatically deleted.

+++ image_manager.cpp
@@ -38,7 +38,10 @@ struct RemovablePath
     RemovablePath(const fs::path& path) : path(path) {}
     ~RemovablePath()
     {
-        fs::remove_all(path);
+        if (fs::exists(path))
+        {
+            fs::remove_all(path);
+        }
     }
 };
